+++ src/main/java/com/github/leandromoraesrj/sincadbackend/domain/InscricaoRegimeApuracao.java
@@ -1,6 +1,7 @@
 package com.github.leandromoraesrj.sincadbackend.domain;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Objects;
 
@@ -25,9 +26,7 @@ public class InscricaoRegimeApuracao implements Serializable {
 	@SequenceGenerator(name = "se_inscricao_estab_reg_apur_generator", sequenceName = "se_inscricao_estab_reg_apur", allocationSize = 1)
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "se_inscricao_estab_reg_apur_generator")
 	@Column(name = "sq_inscricao_estab_reg_apur")
-	private Long id;
-	//sq_inscricao_estabelecimento
-	//sq_regime_apuracao	
+	private BigDecimal id;	
 	@Column(name = "in_excluido")
 	private boolean excluido;
 	@Temporal(TemporalType.DATE)
@@ -46,7 +45,7 @@ public class InscricaoRegimeApuracao implements Serializable {
 	public InscricaoRegimeApuracao() {		
 	}
 
-	public InscricaoRegimeApuracao(Long id, boolean excluido, Date inicio, Date fim, Inscricao inscricao,
+	public InscricaoRegimeApuracao(BigDecimal id, boolean excluido, Date inicio, Date fim, Inscricao inscricao,
 			RegimeApuracao regimeApuracao) {
 		super();
 		this.id = id;
@@ -57,11 +56,11 @@ public class InscricaoRegimeApuracao implements Serializable {
 		this.regimeApuracao = regimeApuracao;
 	}
 
-	public Long getId() {
+	public BigDecimal getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(BigDecimal id) {
 		this.id = id;
 	}
